Convert to TS and format
